rename folders and add healtcheck endpoint
