+++ backend/.gitignore
@@ -1,2 +1,2 @@
 #ignore models
-fernec_api/ia_models/
+fernec/ia_models/
