+++ backend/main.py
@@ -1,12 +1,14 @@
 import uvicorn
 from fastapi import FastAPI
-from fernec_api.predict import router as predict_router
+from fernec.predict import router as predict_router
+from manage.get_state import router as state_router
 
 def get_application() -> FastAPI:
     application = FastAPI(title='FERNEC API')
 
     #Add routers
     application.include_router(predict_router)
+    application.include_router(state_router)
     return application
