+++ backend/shared/models.py
@@ -8,10 +8,3 @@ class StatusResponse(GenericModel, Generic[DataT]):
     success: bool
     message: Optional[str] = None
     data: Optional[DataT] = None
-
-class ImageItem(BaseModel):
-    image_base64: str
-
-class ImagePrediction(BaseModel):
-    predictions: list[float]
-    emotion: str
